Float to Decimal in order to add precision

+++ MarteliveryAPI/Models/DTOs/Admin/AdminQuoteUpdateDTO.cs
@@ -2,8 +2,8 @@
 {
     public class AdminQuoteUpdateDTO
     {
-        public float PricePerKm { get; set; }
-        public float TotalPrice { get; set; }
+        public decimal PricePerKm { get; set; }
+        public decimal TotalPrice { get; set; }
         public string Status { get; set; }
         public string UserId { get; set; }
         public string ParcelId { get; set; }
