+++ MarteliveryAPI/Models/DTOs/Carrier/CarrierQuoteDTO.cs
@@ -2,8 +2,8 @@
 {
     public class CarrierQuoteDTO
     {
-        public float PricePerKm { get; set; }
-        public float TotalPrice { get; set; }
+        public decimal PricePerKm { get; set; }
+        public decimal TotalPrice { get; set; }
         public string ParcelId { get; set; }
     }
 }
